Fix JS/Native CI: make JDKVersion.fromRuntimeApi lazy so its reflection is DCE'd off-JVM (eager val broke linking for any code referencing JDKVersion); move Data numeric int/float/double discrimination tests (fold/render/updateAs) to JVM-only DataValuesJvmSpec (JS/Native box all numbers as Double)
